requirements.txt
alembic==1.4.1
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cssselect==1.1.0
dramatiq==1.8.1
feedgen==0.9.0
Flask==1.1.1
Flask-BabelEx==0.9.4
flask-dramatiq==0.6.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
Mako==1.1.2
MarkupSafe==1.1.1
passlib==1.7.2
prometheus-client==0.7.1
psycopg2-binary==2.8.4
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
readability-lxml==0.7.1
redis==3.4.1
requests==2.23.0
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==1.0.0
WTForms==2.2.1
